+++ CHANGELOG.md
@@ -6,6 +6,10 @@ BC breaking changes:
 
 - the default template is now `default.twig` (it was previously `page.twig`)
 
-New features:
+Major new features:
 
 - [#10](https://github.com/mnapoli/Couscous/pull/10) automatically install Bower dependencies if a `bower.json` file is in `website/public/`
+
+Minor new features:
+
+- The preview command will show the name of the files that have changed when regenerating